requirements.txt
langchain
langchain-openai
langchain-community
langchain_milvus
pymilvus
streamlit
# Unstructured for parsing
pdf2image
pdfminer.six
pillow-heif
pikepdf
pypdf
pytesseract
nougat-ocr
matplotlib
unstructured[pdf]
unstructured_inference
sentence-transformers
#Others
python-dotenv

#CrewAI
grpcio==1.63.0
crewai
crewai[tools]
sentence-transformers

llama-index 
llama-index-vector-stores-milvus 
pymilvus 
llama-index-llms-openai 
llama-index-readers-file